KEG1-52 - Creo primera versión thymeleaf ejercicios

+++ src/main/java/repositories/EjerciciosRepository.java
@@ -4,5 +4,5 @@ import entities.CategoriaEjercicios;
 import entities.EjercicioOpMul;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EjerciciosRepository extends JpaRepository<EjercicioOpMul, Integer> {
+public interface EjerciciosRepository extends JpaRepository<EjercicioOpMul, Long> {
 }
